requirements/requirements-gpu.txt
absl-py==0.1.9
bleach==1.5.0
click==6.7
cycler==0.10.0
decorator==4.2.1
futures==3.1.1
h5py==2.7.1
html5lib==0.9999999
imageio==2.2.0
Keras==2.1.3
Markdown==2.6.11
matplotlib==2.1.2
networkx==2.1
numpy==1.14.0
Pillow==5.0.0
pip-autoremove==0.9.0
pkg-resources==0.0.0
protobuf==3.5.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.13.1
scipy==1.0.0
six==1.11.0
tensorflow-gpu
tensorflow-tensorboard==1.5.0
tqdm==4.19.5
Werkzeug==0.14.1
